requirements.txt
requests==2.32.3
mysql-connector-python==9.0.0
jsonschema==4.23.0
python-dotenv==1.0.1
pytest==8.3.2